+++ controller_manager/src/ros2_control_node.cpp
@@ -45,13 +45,17 @@ int main(int argc, char ** argv)
   auto cm = std::make_shared<controller_manager::ControllerManager>(executor, manager_node_name);
 
   RCLCPP_INFO(cm->get_logger(), "update rate is %d Hz", cm->get_update_rate());
+  const int thread_priority = cm->get_parameter_or<int>("thread_priority", kSchedPriority);
+  RCLCPP_INFO(
+    cm->get_logger(), "Spawning %s RT thread with scheduler priority: %d", cm->get_name(),
+    thread_priority);
 
   std::thread cm_thread(
-    [cm]()
+    [cm, thread_priority]()
     {
       if (realtime_tools::has_realtime_kernel())
       {
-        if (!realtime_tools::configure_sched_fifo(kSchedPriority))
+        if (!realtime_tools::configure_sched_fifo(thread_priority))
         {
           RCLCPP_WARN(
             cm->get_logger(),